3-9 times faster again, now 100ms for `[[[[NewType(123)]]]]`, 15ms for `NewType(123)`


test script:

```julia
struct A x end
struct B x end
struct C x end
struct D x end

for m in ["Starting", "Second time"]
    @info m

    io = IOContext(IOBuffer())

    @time PlutoRunner.format_output([[[[A(123)]]]])
    @time PlutoRunner.format_output([[[[B(123)]]]])
    @time PlutoRunner.format_output(C(123))
    @time PlutoRunner.format_output(D(123))

    if false
        asdfasdf(:(struct A end))
    end
end
```


On this branch:

```
[ Info: Starting
  0.354511 seconds (2.20 M allocations: 117.306 MiB, 3.64% gc time)
  0.104522 seconds (352.83 k allocations: 20.238 MiB)
  0.016901 seconds (45.36 k allocations: 2.583 MiB)
  0.016694 seconds (45.36 k allocations: 2.585 MiB)
[ Info: Second time
  0.000550 seconds (780 allocations: 36.672 KiB)
  0.000570 seconds (780 allocations: 36.688 KiB)
  0.000203 seconds (278 allocations: 12.406 KiB)
  0.000187 seconds (278 allocations: 12.406 KiB)
```

on master:

```
[ Info: Warmup
[ Info: Starting
  0.506999 seconds (2.54 M allocations: 134.934 MiB, 2.40% gc time)
  0.270330 seconds (848.05 k allocations: 47.018 MiB, 1.34% gc time)
  0.094456 seconds (297.32 k allocations: 16.275 MiB, 3.88% gc time)
  0.090100 seconds (297.33 k allocations: 16.278 MiB)
[ Info: Second time
  0.000593 seconds (926 allocations: 39.297 KiB)
  0.000596 seconds (926 allocations: 39.266 KiB)
  0.000206 seconds (344 allocations: 13.703 KiB)
  0.000221 seconds (344 allocations: 13.703 KiB)
```

on 0.12.13:

```
[ Info: Starting
  3.559646 seconds (16.29 M allocations: 851.412 MiB, 4.61% gc time)
  3.175303 seconds (14.09 M allocations: 736.600 MiB, 3.10% gc time)
  0.100442 seconds (297.55 k allocations: 16.211 MiB, 4.98% gc time)
  0.093871 seconds (297.56 k allocations: 16.212 MiB)
[ Info: Second time
  0.000624 seconds (1.01 k allocations: 48.125 KiB)
  0.000563 seconds (1.01 k allocations: 48.125 KiB)
  0.000199 seconds (353 allocations: 15.344 KiB)
  0.000282 seconds (353 allocations: 15.344 KiB)
```